+++ packages/highlight/src/language/custom/javascript/index.ts
@@ -12,7 +12,7 @@ const jsx = new HighlightCustomLanguage("jsx", [], {
   grammar: "jsx",
 });
 
-javascript.replaceToken(
+javascript.replaceTokenRule(
   "keyword",
   "control-flow",
   (oldToken) => {
@@ -25,13 +25,17 @@ javascript.replaceToken(
   }
 );
 
-jsx.replaceToken("keyword", "control-flow", (oldToken) => {
-  return {
-    ...oldToken,
-    pattern:
-      /\b(?<!\.)(?:(await(?= |\()|break(?=\b)|catch(?=[\s]*\()|continue(?=\b)|do(?=\b)|else(?=\b)|finally(?=\b)|for(?=\b)|if(?=\b)|return(?=\b)|switch(?=\b)|throw(?=\b)|try(?=\b)|while(?=\b)|yield(?=\b)))/,
-    alias: "control-flow",
-  };
-});
+jsx.replaceTokenRule(
+  "keyword",
+  "control-flow",
+  (oldToken) => {
+    return {
+      ...oldToken,
+      pattern:
+        /\b(?<!\.)(?:(await(?= |\()|break(?=\b)|catch(?=[\s]*\()|continue(?=\b)|do(?=\b)|else(?=\b)|finally(?=\b)|for(?=\b)|if(?=\b)|return(?=\b)|switch(?=\b)|throw(?=\b)|try(?=\b)|while(?=\b)|yield(?=\b)))/,
+      alias: "control-flow",
+    };
+  }
+);
 
 export { javascript, jsx };